ID: 3488008 Eyeform
- Use the following flag to enable the eyeform: new_eyeform_enabled =
true
- Eyeform opens from "EF" link on Appointment or Search screens
- If opened from an appointment:
	* The "Reason" field (on the appointment) is used to populate the
"Today's Concern" field (top of the eyeform)
	* The "Today's Concern" field updates the "Reason" field on the
appointment when edited
	* Click on the "Today's Concern" to edit; hit enter or click on
something else to finish editing

- Highlight
	* By default, when opening the eyeform all data from the last 3 hours
is automatically highlighted
	* Move the slider at the top of the page left and right to change
highlighting to different periods
	* Data that was recorded during THIS appointment should be BOLD and in
RED (and highlighted if selected)

- Clicking on any of the headers of the grey boxes (with a popup icon in
their title) opens the history or full view for that type.
- Left bar boxes
	* Clicking on an item (on the left bar) will open that item in a new
window, except for these exceptions: Specs History, Procedures, and
Tickler.
	* If there are more than 5 items in any box, the additional items will
be hidden until you click "Show All"
- Specs History/Procedures/Tickler
	* Clicking on "+" brings up an inline popup editor to create a new item
for any of these types
	* Clicking on any of the elements inside of these boxes brings up an
inline editor or viewer for the data; any changes to that data can be
made there
- Diagrams/EForms
	* Clicking on "+" shows 4 diagrams (Retina, Eye Saggital, Eye Lid/NLD,
Ant Segment) and the Rich Text Letter, ONLY if those eforms are
installed on your OSCAR.

- CPP Boxes
	* Diagnostics, Patient Log, Notes: clicking on the "+" icon will create
a new item and place the text cursor inside.
	* Diagnostics, Patient Log, Notes, Previous Plans: clicking on any
existing item will allow the user to edit that item.  Click on a
different part of the page or press enter to save.
	* Diagnostics, Patient Log, Notes, Previous Plans: hover over any
existing item and click "Archive" to archive that specific note.
	* Diagnostics, Patient Log, Notes, Previous Plans, Impression History:
hover over the underlined date to see the name of the provider who
recorded the note.
	* Diagnostics, Patient Log, Notes, Previous Plans, Past Medical
History: click on the header to show the entire history of all notes in
that box.

- Measurements
	* Click "Modify" to open the editor for measurements.
	* When measurements are modified and the Measurements box is closed,
the measurements are saved.
	* Click on any previous measurements item to show the measurements as
they were recorded at that time.
		* You can't edit the measurements displayed here (there's an
explanation given at the top).
	* Click on the "Next" or "Previous" buttons on the Measurements box
(next to "Modify") to highlight the next (older) or previous (newer)
item that was recorded.  The box's scroll bar will jump to the recorded
item and the item will be given a faint grey background.

- Impression History
	* Clicking on any of the previously recorded impressions will invoke
the eyeform PDF creation routine and send the result to the user.

- Current Presenting Issue, Impression, Plan
	* These boxes grow to always fit all of the text that you type inside
of them.
    * Copy Appointment Reason: copies the contents of the appointment
reason/"Today's Concern" to the Current Presenting Issue box.
    * History (Current Presenting Issue): displays the OSCAR note
history of the current presenting issue.
	* Copy Last Impression: copies the previous impression (top of the
Impression History list) to the Impression box.
	* No Change: adds "Patient's condition has not changed." to the
Impression box.
	* Stable: adds "Patient's condition is stable." to the Impression box.
	* If there is content in the Impression box already, clicking on any of
the above buttons will add two new lines before appending the text.

- Plan
	* When you click on Sign, Save & Exit, the content of this field is
sent to all users who have the Receptionist role as a tickler.

- Save & Exit
	* Saves all data (Current Presenting Issue, Impression, Plan).  DOES
NOT SIGN THE CURRENT PRESENTING ISSUE, IMPRESSION OR PLAN ITEMS.
	* Does not send the text of the Plan box to the Receptionists.
	* When another user opens the eyeform, the unsigned Current Presenting
Issue, Impression and Plan will be displayed in the appropriate boxes.
	* Closes the window.

- Sign, Save & Exit
	* Saves all data (Current Presenting Issue, Impression, Plan) AND signs
those items.
	* Sends the Plan tickler.
	* Closes the window.

- Sign, Save & Bill
    * Same as Sign, Save & Exit but instead sends the user to the
billing page for the patient.

- Macros
	* "M" button opens the Macro editor.
		* Click "New" to create a new macro; all the fields in the editor are
cleared.
		* Choose a name for the macro (it needs to be unique)
		* Enter an impression note or click "Copy Last Impression"
		* Enter a plan note (or leave empty for no plan)
		* Billing: add billing codes to the list of items that will be billed.
 KNOWN ISSUE: This doesn't display anywhere right now.
		* Save: saves the macro currently being edited.
		* Copy: Copies the selected macro (list on the left hand side of the
editor) and clears the name field.
		* List of macros (left hand side): Click on a macro here to edit.
	* Start typing a macro name to bring up an auto-complete list for all
macros
		* Click on the desired macro or use the arrow keys to select and press
enter to execute.
			* The impression and plan boxes are outlined in red to show that the
macro has modified their values.

Change-Id: I1a03afb00235d64c4855e6f0cab016c7aab3f776