ID: 361257 Rich Text Letter Support for FF18+ restored
Handles ionMonkey's loss of object detection for window[editorname] by
reverting to user agent parsing which is retro compatible

However the method is potentially unreliable as it depends on browsers
detected as IE to support window[editorname] and all others to be more
or less standard compliant.

Tested in FF20 and Chromium 25.

Change-Id: I479ba8ac1d8e9985623d0a3e3b98a4fddf1eed51