ID: 3604601 - Bug Fix -Diabetes flowsheet, cannot enter weight greater than 100
UPDATED - insert err WT in bmi mtype into measurementTypeDeleted
then delete the WT in BMI mtype.

When entering a weight within the diabetes flow sheet over 100 kg it
notifies me that I am only able to input weight between 0 and 100...

it seems the flowsheet is confused between the WT KG type with the WT
BMI and therefore specifying that only a max of 100 is allowed.

https://sourceforge.net/tracker/?func=detail&atid=515432&aid=3604601&group_id=66701

Change-Id: I8677d9ff17312ac85fa8da60aa7ff24a45ce7b4e