2423 - Click Save in Encounter Note Will Cause Next Edit to Be Lost
So since there's currently a hook on the onclose for the window object
to call this onClosing method in newCasemanagementView.js.jsp, I decided
to atleast fix that behavior as per the discussion on the sourceforge
bug.

this change of logic doesn't case about tmpSaveNeeded variable always being true
it's fine if it's not, but if there's a change in the obsdate or the note itself,
we resave it onclose.