Allows HRM reports to be attached to consultations and fixes OLIS print
Allows HRM reports to be attached to consultations, their colour is
currently set to red in the list on the ConsultationFormRequest page.
At this time the ability to print them with the consultation is not
implemented, they must be printed separately.

Fixes printing OLIS labs with the consultation, they no longer give an
error and use the OLIS printing functionality