Adds appointment counter to schedule and two preferences
Adds an appointment counter to the schedule page beside each doctor's
name.  This counts the number of appointments each doctor has that day,
excluding cancelled appointments and no shows.

Adds two new user preferences, the first is a twelve hour format
preferences to display the time in am/pm instead of 24 hours (Does not
change the time on the edit appointment screen).  The second preference
is a shortcut on appointments to the patient's label printing.