Changes the appointment history screens to display status and notes
Changes the add appointment page's appointment history to display the
status of past appointments and to display notes instead of comments.

Changes the demographic appointment history page to display notes
instead of comments.

Adds a max length of 255 characters to the comment textarea on the add
and edit appointment pages