Enhances search to add a referring doctor on new and existing patients
When adding or changing a referral doctor on a new or existing patient,
adds a 'Search Name' link to search by the name of the referral doctor.
If only one match is found for the input name or number, then it will
automatically fill out the referral doctor name and the referral number
on the master file page.  If more than one match is found then it
displays all matches.  Removes the specialty column and adds an address
column.