#4279 NPE when professionalSpecliasts.referralNo is null
When there is a row in the profressionalSpecialists table that has
referralNo = null, it causes a null pointer exception. Changing the null
to the empty string is a quick workaround, but then all matches with the
empty string are pre-checked.
null referralNo should not cause an NPE
speclialists without a referralNo should not appear in the list and
should certainly not be pre-checked.


Change-Id: I8d12a77a1506a4ed0257c98e59e2a97edc513a53