OSCAR-267: Used global replace to replace all slashes when saving
When date is manually entered with slashes ex. xxxx/xx/xx it would
replace all the slashes with hyphens (-).  The replace only replaced one
slash in the string rather than all of them, which led to an invalid
date format.  Used a global replace regex pattern to replace all slashes
in the date with hyphens.