OSCAR-479: Ability to Set the Default City when Adding a new Demographic
Added new property in properties file called default_city, City will now default to the value specified for that property in the properties file.
Made the default city Toronto in the oscar_mcmaster.properties, that way there will be a defult city incase the one in the oscar.properties is not set.