OSCAR-482: Labels Too Large
Address, City, Province and Postal Code line has been split into 2 separate lines, 1 for the address and 1 for the City, Province and Postal Code. In addition to this, y values of elements have been modified in order to fit on to label without using fit-to-page in print dialog. Bolded and underlined demographic name and chart # has been reduced in size to 9px, since it is bolded and underlined difference cannot really be seen.