OSCAR-447: Always show "due" preventions at the top of the Preventions list in the echart
Added a new ArrayList for duePreventions. Items are added to the duePreventions arraylist by checking colour for red and checking if the div string is preventions. This is done this way because lab results can sometimes have a red colour.
Due preventions are displayed right at the top followed by preventions with no dates.