OSCAR-479: Ability to Set the Default City when Adding a new Demographic
Refined logic, if the Province is Ontario and the billingCentre is Toronto, the default city is set to Toronto.
If the billing centre is not Toronto but the province is Ontario, the default city set in the properties file is used.
If both conditions are not satisfied, the default city is left blank.