OSCAR-520: Date Validation for Observation Date on Document Preview (from Inbox) Allows for Non Existent Date
Observation Date was not checking for leap year. If a leap year occurs, checks if month is February and if the day is greater than 29. If a leap year doesn't exist checks the month for February, if the month is February it checks if the day is greater than or equal to 29. Alert message notifies user that date does not exist.