OSCAR-481: Error when Saving a Hospital Billing When Multi Site Enabled
It was trying to get the index of a "|" that did not exist in the cases that I have tested. Usually the "|" separates the ProviderNo and OHIPNo. Made a check to see if the string contains a "|" if it does contain "|" it will use the original set method, otherwise it will just set the providerNo the same way it sets it when Multi Site is not enabled.