OSCAR-529: After saving a new external contact under Manage Contacts, redirect to the Search Contacts page
Redirected the contactSearch page 'Add Contact' button directly to addContact page, as I also redirected the struts-xml to direct to contactSearch instead of back to addEditContact when submitting a newly added contact. Threw parameter values through the contactAction.java with parameter action forward so they would carry over from addEditContact. contactSearch.jsp was only searching on form submission as that's how it was designed, so I removed those statements, also keyword value was very strangely always throwing a NullPointerException when entered into the function but no where else, so replaced it and the hidden input with elementName, set it to only display if its an actual added value and not contact_2.contactName, and move the function below the form so it works properly.