OSCAR-607: Deleting Physician Name from Inbox "Search" Still Searches All Providers
Provider No was stored in a hidden field and was not changed when the textbox information for the provider name was changed. Made an onchange event that will trigger when the textbox is changed. The providerNo string will then be changed to "0". When a different provider is selected from the search it uses the information from the provider that was selected.