OSCAR-622: Ability to File HRM Reports from "List View" of the Inbox by Checking the Checkbox(s)
Original signoff functionality did not account for HRM Documents. Added a check to see if the checked checkbox is a type of HRM. If the type is HRM a separate query string is made to handle the functionality for signing off HRM's. The function will then submit an Ajax request only if the query string for the HRM Documents is not empty. The request will then submit to the signOff function in the HRMModifyDocumentAction.java file. The signoff method has been updated to account for multiple reportIds by using getParameterValues instead of getParameter. It will then loop through the reportIds and try to sign off theHRM document.