OSCAR-659: Unable to Type in the "Name" Field of the Manage Contacts Area of Master File Until the User Clicks "Search" (Read Only Field)
Removed readonly attribute from input tag, added functionality to automatically search the value that is in the newly write-able input box when 'search' is clicked. (contactId is hardcoded to 0 always, no idea why it's used. Also, my build has never had functionality to add a contact by clicking their name as a search result, but it works on the pilot, and this fix seems to have no effect on that functionality).