OSCAR-677: Display Cell Phone Number on Consultation RequestsDisplay Cell Phone Number on Consultation Requests
Decided to switch the search function from findAll(bool) to the already existing findAllOrderByLastName(), and add a check for inactive providers, like findAll(bool). I added this check because the only other usage of findAllOrderByLastName() was in providerPrivilege.jsp, and it would not effect the results if it were to only search for active providers, not to mention the original findAll(bool) had a hardcoded false being passed to it, so it was almost pointless to do the check. But in the end, it was easier to make a slight modification to the findAllOrderByLastName function than to tamper with the query in findAll(bool), seeing as that as FAR more usages in the project, and fixes the problem rather simply.