OSCAR-706: Preference to Change Time from Military to Standard Should Also Affect "Make an Appt" Screen
Changed the time being displayed from the actual start_time parameter to a formatted version of it depending on whether the 12-hour format was set in preferences. Changed stime JS variable in calculation from the start_time input tag value to the parameter as the input tag is not formatted and will mess up calculation. Formatted start_time/end_time parameter back to 24-hour time in linked files so as to not interfere with calculations for duration,etc. and the display in any of these files has also been formatted accordingly.