OSCAR-752: A1C Value Mapping Issue on MedHealth Labs
Added an additional parameter to the querying of the mapping table. Before, it was finding the value associated with the loinc code and type (in this case, GENERAL CHEMISTRY), and then setting variables to whatever the name (in this case, A1c) values were in the mappingType table. It would do this for every measurement in the lab, as there was no way to check and make sure that it was the name associated with that loinc code that was mapped. Adding the name parameter allowed the query to not only ensure the loinc code existed and was mapped to a type (GENERAL CHEMISTRY), but ensure the name of the measurement that was being read from the lab file matched the name associated with that loinc code/type. (Tested mapping each measurement type in the lab file to general chemistry, worked for all)