OSCAR-753: E082A Service Code Not Adding 30% Fee Properly
Not sure what the Min and Max settings for billing % are in admin, but in the case of this file they are referenced from Java variables here, and this logic will always set the calculated sub-total amount to the Min value, as it stores 0 by default. Even if the value ($10.41 in this case), were to be in the textbox, these checks would be pointless. Left the Java variables above for now in case they need to be properly used in the future (vecPercMin and vecPercMax).