OSCAR-772: Error Finding HRM Documents Attached To Consultation
Noticed this when looking into Brent's OSCAR-770 Log Error bug. Taking the null value and setting it to 'null' in a javascript variable, then passing that through to the backend, which cannot be converted to a number. Just added a condition to only try the query if the consultationId doesn't equal "null".