OSCAR-774: Open Appointment Search Not Working With Provider Schedule Template
Was printing all days no matter what with the logic. Created a function to query the rschedule table for Active schedules with the current provider, and added a condition to ensure the provider has an active template, if they do, return nothing for days that are not part of the schedule. If they do not have a template, return all days as available.