OSCAR-593: Ability to Delete External Providers Without Causing Error on Previously Completed Consultation Requests
Added the ability to delete a provider without eliminating their record in the back end. There was a previously commented out button which would link to an action, completely removing the record from the database. Just added a deleted field to the professionalSpecialist table, and had the button set it to true and merge the changes. Then for display, it would only display providers for services and for editing if their deleted field was set to false. Added update-2017-05-01.sql file with sql line used, and modified oscarinit to add deleted field to the professionalSpecialists table.