OSCAR-786: Sign off Button on Preview Mode for HRM Result Closes back to List View, Should go to Next Report
Added containing div tag with the HRM Document ID appended in the id, used to pass into similar function for documents that I added to HRMs after they are merged with the database. Function fades it and removes from the list, displaying the next HRM as opposed to closing the window all together.