OSCAR-889: Adding Family Doctor to Appointment Card
Added family doctor to appointment card, spoke with Ashley and she said it would be good to also have the referral doctor on there. Also, I'm not sure why, the but getFamilyDoctor function of the demographic class pulls the value from the Referral doctor field on the patients master file, and the getFamilyPhysician pulls from the family doctor field. Ashley said this makes sense for the family doctor pulling from referral doctor field, so I just went with that. Can be reverted if that's not proper.