OSCAR-899: Waiting List Popup Ask if the Patient Should be Removed When Booking Appointment - Not Automatically Remove
Was using a custom modal originally, as Javascript Alert does not allowed the modification of the 'Ok' and 'Cancel'. Only problem was the custom modal doesnt freeze the pages execution until a button is clicked on the Alert, so the 'fake alert' closes instantaneously with the submission of the page. Sara ended up saying to just remove the pop up functionality so they can't remove patients from the waiting list upon appointment creation (left the function in case it wants to be implemented again in the future).