OSCAR-797: Outward Facing Fax Number Not Consistent with the Telephone Format
Modified the fax number from the fax config section directly in the PDF creation; formatting it within the Configure Fax section of the Admin section isn't possible due to the length of the field where it is stored being too small to store a number in the format of ###-###-#### (12 length). (Also formatted clinic phone and clinic fax IF they don't contain '-', since they can be added without that formatting as well.