OSCAR-923: Next Appointment Search - Appt Type not sorting
Spoke with Sara, removed the changes requested from OSCAR-773, temporarily set it so only Lunch and Vacation will not show as available. Going to write a ticket and implement capability to set newly created template codes as 'Available' or 'Not Available', to avoid conflict with differing clinics wanting different codes/custom codes to be available or not.