OSCAR-926: Ability to Set Template Codes as Available/Not Available
Added ability to set existing and custom template codes to be marked as Available/Unavailable when completing an appointment search from the schedule screen. Fixed being able to save without editing a code, which was causing an ArrayIndexOutOfBounds exception, also fixed the fact that it was grabbing the first character in the order for the same letter (A and a, M and m) by switching to a native query (JPA doesn't support binary) and returning setting the condition to the binary of the character.