Add old functionality of original Util class
This is a stop-gap retrieval to provide a very limited feature related
to prevention detection. New version (Util) assumes user information is
available (logged in user).  The PCDS process is run as a job so logged
in user information is not available.

There should be a way to use the new class but that is for another day.