OSCAR-1321: System Management \> Rx Settings does not change for all users
Modified the rxManage section so that it will be a generally applied to all providers setting, as opposed to provider specific. Spoke with Sara and she said if we get enough request to make MRP display on Rx Printouts a preference, we can implement that at a later date.

(cherry picked from commit c274f274b26428929d3878a1ba2f8bfe9c6794f0)