OSCAR-1446: Pop Up Notification for Clients from KAI
Created a consent popup notification that will inform clients of the new modules not being open source. Arjun said he is going to work on the message and link this week, so for now I just have a placeholder message and link to google.ca, will update the branch and squash the commits when I receive the information.

(cherry picked from commit 6cb0956)