OSCAR-1720: Manually Setting Status to "Confirmed" will not Send Reminder
Added the ability to stop all associated appointment reminders from going out by updating the status to that which was set in the appointment reminder settings. Must migrate the table in so as not to effect clients without the online booking/appointment reminders.

(cherry picked from commit 68c6c0d)