OSCAR-2410: Hide Mailing Address Unless Populated (Checkbox to Enable)
Added a checkbox in both masteredit and new demographic that hides or shows the mailing address information. By default, the checkbox is off and the mailing address information is hidden, unless it was already populated. If it's populated, it will display as normal.

Once the checkbox is click (either off or on) that selection is saved and instead of checking whether or not the forms were populated, it will use that selection instead.