OSCAR-1218: Autoloading of PDF's When HRM Sent as HTML Link
Checks for PDFs and embeds them into the page similar to how it does with images. The download file action was modified to account for PDFs as well. Displays a message if the PDF couldn't be displayed in which case they may need to update chrome or disable PDF downloading.