OSCAR-1553: Missing Server Side Validation
Added backend validation to areas within Admin that lacked it to prevent invalid values being sent back into the database. This includes ensuring some values are numeric, existing, and within the bounds of the columns maximum length, before they are added to the database. Since most of these validations can only be invalid if somebody is trying to do something malicious, most of the time the user will just see a blank page upon sending the values, and their input becomes ignored.