OSCAR-2815 Release Issue - Fixed signature stamp not saving to backend
When a provider different from the prescribing physician logs in to reprint an Rx, the signature stamp of the prescribing physician is missing. This was because the signature stamp was not saving to the backend. This fix ensures the stamp is saved to the digistal_signature table and the signatureID exists in the prescription table for that stamp.